refactor: main controller 생성 및 메인페이지 구현
